rename reticulum path store module
