+++ crates/libs/rns-transport/src/transport/mod.rs
@@ -64,7 +64,7 @@ mod link_table;
 mod packet_cache;
 mod path_requests;
 mod path_table;
-mod persistence;
+mod reticulum_path_store;
 mod tunnels;
 
 pub mod test_bridge {
